MIL/requirements.txt
torch
torchvision
dgl
h5py
matplotlib
networkx
opencv-python
pandas
Pillow
PyYAML
scikit-image
scikit-learn
scipy
seaborn
tqdm
sdv
polars
timm
hiyapyco
openTSNE
openpyxl
umap-learn
prettytable